fixes for clang checks.

+++ src/examples/libpmemobj/libart/arttree.c
@@ -157,7 +157,7 @@ initialize_context(struct ds_context *ctx, int ac, char *av[])
 				}
 				break;
 			case 'n': {
-				int insertions;
+				long int insertions;
 				insertions = strtol(optarg, NULL, 0);
 				if (insertions != LONG_MIN &&
 				    insertions != LONG_MAX) {
